Move dynamic compile to NetDaemon.AppModel.SourceDeployedApps.csproj
