+++ src/AppModel/NetDaemon.AppModel.SourceDeployedApps/Compiler/CompileSettings.cs
