+++ src/AppModel/NetDaemon.AppModel.SourceDeployedApps/Compiler/ICompiler.cs
