+++ src/AppModel/NetDaemon.AppModel.SourceDeployedApps/Compiler/ISyntaxTreeResolver.cs
