+++ src/AppModel/NetDaemon.AppModel.SourceDeployedApps/Compiler/SyntaxTreeResolver.cs
@@ -1,6 +1,7 @@
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.Text;
+using Microsoft.Extensions.Options;
 
 namespace NetDaemon.AppModel.Internal.Compiler;
