+++ src/AppModel/NetDaemon.AppModel.SourceDeployedApps/GlobalUsings.cs
@@ -1,0 +1,8 @@
+global using System;
+global using System.Collections.Generic;
+global using System.Runtime.CompilerServices;
+global using System.Text;
+
+// Make the internal visible to test project
+[assembly: InternalsVisibleTo("NetDaemon.AppModel.Tests")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
