+++ src/AppModel/NetDaemon.AppModel/Internal/AppAssemblyProviders/IAppAssemblyProvider.cs
@@ -2,7 +2,7 @@ using System.Reflection;
 
 namespace NetDaemon.AppModel.Internal.AppAssemblyProviders;
 
-internal interface IAppAssemblyProvider
+public interface IAppAssemblyProvider
 {
     public Assembly GetAppAssembly();
 }
